requirements/test.txt
asynctest==0.13.0
codecov==2.1.11
coverage==5.3.1
pytest==6.2.1
pytest-aiohttp==0.3.0
pytest-cov==2.10.1
pytest-mock==3.4.0
